+++ src/set_svg_matrix2d_b.ts
@@ -1,12 +1,11 @@
 import { B, be_ } from '@ctx-core/object'
-import { get } from '@ctx-core/store'
-import { svg_margin_T, svg_margin$_b } from './svg_margin$_b.js'
-import { svg_width$_b } from './svg_width$_b.js'
-import { svg_height$_b } from './svg_height$_b.js'
-import { svg_content_width$_b } from './svg_content_width$_b.js'
-import { svg_content_height$_b } from './svg_content_height$_b.js'
 import type { Placement } from './Placement.js'
-import type { svg_Ctx } from './svg_Ctx'
+import { svg_content_height$_b } from './svg_content_height$_b.js'
+import { svg_content_width$_b } from './svg_content_width$_b.js'
+import { svg_height$_b } from './svg_height$_b.js'
+import { svg_margin$_b } from './svg_margin$_b.js'
+import { svg_width$_b } from './svg_width$_b.js'
+import type { svg_Ctx } from './svg_Ctx.js'
 const key = 'set_svg_matrix2d'
 export const set_svg_matrix2d_b:B<svg_Ctx, typeof key> = be_(key, ctx=>{
 	return set_svg_matrix2d
@@ -15,15 +14,15 @@ export const set_svg_matrix2d_b:B<svg_Ctx, typeof key> = be_(key, ctx=>{
 	) {
 		const svg_margin =
 			opts.svg_margin
-			|| get(svg_margin$_b(ctx))
-			|| { top: 20, right: 20, bottom: 60, left: 100 } as svg_margin_T
+			|| svg_margin$_b(ctx).$
+			|| { top: 20, right: 20, bottom: 60, left: 100 } as Placement
 		const svg_width =
 			opts.svg_width
-			|| get(svg_width$_b(ctx))
+			|| svg_width$_b(ctx).$
 			|| NaN
 		const svg_height =
 			opts.svg_height
-			|| get(svg_height$_b(ctx))
+			|| svg_height$_b(ctx).$
 			|| NaN
 		const {
 			left,
@@ -31,13 +30,13 @@ export const set_svg_matrix2d_b:B<svg_Ctx, typeof key> = be_(key, ctx=>{
 			top,
 			bottom,
 		} = svg_margin
-		const svg_content_width = svg_width - left - right
 		const svg_content_height = svg_height - top - bottom
-		svg_margin$_b(ctx).set(svg_margin)
-		svg_width$_b(ctx).set(svg_width)
-		svg_height$_b(ctx).set(svg_height)
-		svg_content_width$_b(ctx).set(svg_content_width)
-		svg_content_height$_b(ctx).set(svg_content_height)
+		const svg_content_width = svg_width - left - right
+		svg_margin$_b(ctx).$ = svg_margin
+		svg_height$_b(ctx).$ = svg_height
+		svg_width$_b(ctx).$ = svg_width
+		svg_content_height$_b(ctx).$ = svg_content_height
+		svg_content_width$_b(ctx).$ = svg_content_width
 	}
 })
 export interface svg_matrix2d_set_opts_T {