+++ src/svg_matrix2d_ctx$_b.ts
@@ -1,11 +1,12 @@
 import { B, be_ } from '@ctx-core/object'
 import { derived, Readable } from '@ctx-core/store'
-import { maybe_svg_margin_T, svg_margin$_b } from './svg_margin$_b.js'
-import { maybe_svg_width_T, svg_width$_b } from './svg_width$_b.js'
-import { maybe_svg_height_T, svg_height$_b } from './svg_height$_b.js'
-import { maybe_svg_content_width_T, svg_content_width$_b } from './svg_content_width$_b.js'
-import { maybe_svg_content_height_T, svg_content_height$_b } from './svg_content_height$_b.js'
-import type { svg_Ctx } from './svg_Ctx'
+import { svg_margin$_b } from './svg_margin$_b.js'
+import { svg_width$_b } from './svg_width$_b.js'
+import { svg_height$_b } from './svg_height$_b.js'
+import { svg_content_width$_b } from './svg_content_width$_b.js'
+import { svg_content_height$_b } from './svg_content_height$_b.js'
+import type { svg_Ctx } from './svg_Ctx.js'
+import type { Placement } from './Placement'
 const key = 'svg_matrix2d_ctx$'
 export const svg_matrix2d_ctx$_b:B<svg_Ctx, typeof key> = be_(key, ctx=>
 	derived([
@@ -31,13 +32,13 @@ export const svg_matrix2d_ctx$_b:B<svg_Ctx, typeof key> = be_(key, ctx=>
 	) as svg_matrix2d_ctx$_T
 )
 export interface svg_matrix2d_ctx_T {
-	svg_margin:maybe_svg_margin_T
-	svg_width:maybe_svg_width_T
-	svg_height:maybe_svg_height_T
-	svg_content_width:maybe_svg_content_width_T
-	svg_content_height:maybe_svg_content_height_T
+	svg_margin:Placement|undefined
+	svg_width:number|undefined
+	svg_height:number|undefined
+	svg_content_width:number|undefined
+	svg_content_height:number|undefined
 }
-export interface svg_matrix2d_ctx$_T extends Readable<svg_matrix2d_ctx_T> {}
+export type svg_matrix2d_ctx$_T = Readable<svg_matrix2d_ctx_T>
 export {
 	svg_matrix2d_ctx$_b as b__ctx__matrix2d__svg,
 }