+++ src/svg_preprocess_.ts
@@ -1,9 +1,7 @@
-import { markup_, markup_T } from './markup_.js'
-export function svg_preprocess_(builder_opts = {}):svg_preprocess_T {
+import { markup_, markup_builder_opts__T, markup_T } from './markup_.js'
+export function svg_preprocess_(builder_opts:markup_builder_opts__T = {}):svg_preprocess_T {
 	const markup = markup_(builder_opts)
-	return {
-		markup,
-	}
+	return { markup, }
 }
 export {
 	svg_preprocess_ as _svg_preprocess,